Pulled out some unnecissary features
Removed the augmented Feed generating class.

Removed the predict_with_saved_input function which would store the pre-activation inputs for each node in an auxiliary feed while computing the activations..

Moved test_GD to test_xor, since it tests an XOR pattern
Moved test_regression to test_sin, since it tests a sin pattern, and is only working on a discretized sin wave in any case.
Added an image, which I hope I can get into the README

+++ test/test_sin.cpp
@@ -37,7 +37,7 @@ typename Feed::Output float2label(float f) {
 float label2float(const Feed::Output& layer) {
   int i = max_element(layer.begin(), layer.end()) - layer.begin();
   cout << "Index: " << i << " yielding " << ((i * 2) / (float)num_bins) - 1 << endl;
-  return (i * 2) / num_bins - 1;
+  return (float)(i * 2) / (float)num_bins - 1;
 }
 
 
@@ -72,7 +72,7 @@ int main() {
     ofstream file("learned-sin.txt");
     for(float x = - 2 * M_PI
 	  ; x < (2 * M_PI)
-	  ; x += (M_PI / 4)) {
+	  ; x += (M_PI / 16)) {
       feed.layer[0] = x;
       // print_array(feed.output_layer()) << endl;
